+++ apps/erlangbridge/test/lsp_syntax_SUITE_data/include_with_error.erl
@@ -1,0 +1,5 @@
+-module(include_with_error).
+-include("broken_include.hrl").
+-export([go/0]).
+
+go() -> ok.