+++ apps/erlangbridge/test/lsp_syntax_SUITE_data/include_with_error_comment.erl
@@ -1,0 +1,6 @@
+-module(include_with_error_comment).
+
+-include("broken_include.hrl"). % trailing comment
+-export([go/0]).
+
+go() -> ok.