+++ apps/erlangbridge/test/lsp_syntax_SUITE_data/include_with_error_crlf.erl
@@ -1,0 +1,6 @@
+-module(include_with_error_crlf).
+
+-include("broken_include.hrl").
+-export([go/0]).
+
+go() -> ok.